requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
pdfplumber==0.11.4
anthropic==0.34.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
